apps/backend/pyproject.toml
[tool.poetry]
name = "lms-backend"
version = "0.1.0"
description = "Backend for the Learning Management System"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 